HealthCareforReal/Backend/requirements.txt
fastapi
mariadb
uvicorn[standard]
bcrypt
passlib[bcrypt]
pydantic[email]
pydantic-settings
python-jose[cryptography]
python-multipart
httpx
python-dotenv
aiofiles
requests
sqlmodel